# Consolidate flushes to reduce overhead