This change batches calls to `flush` together where possible. `flush` is a relatively expensive call. By batching calls to `flush` together, we can sell out write latency (which isn't really a primary concern) to pick up some gains in overall throughput (which _is_ a primary concern). Some benchmarks with a local client and server show a significant improvement with this change in place:

### Before

| Concurrent connections | Time to send 10,000 notifications |
|------------------------|-----------------------------------|
|                      1 |             0.336 ± 0.003 seconds |
|                      4 |             0.231 ± 0.019 seconds |
|                      8 |             0.221 ± 0.013 seconds |

### After

| Concurrent connections | Time to send 10,000 notifications |
|------------------------|-----------------------------------|
|                      1 |             0.312 ± 0.002 seconds |
|                      4 |             0.160 ± 0.009 seconds |
|                      8 |             0.152 ± 0.013 seconds |

This closes #655.