@jchambers 2018-12-26
Yow. Combined with #658, this makes a _huge_ difference in throughput for local benchmarks:

### Pushy v0.13.6

| Concurrent connections | Time to send 10,000 notifications |
|------------------------|-----------------------------------|
|                      1 |             0.333 ± 0.002 seconds |
|                      4 |             0.251 ± 0.010 seconds |
|                      8 |             0.212 ± 0.018 seconds |

### After updating to Netty 4.1.32 and consolidating flushes

| Concurrent connections | Time to send 10,000 notifications |
|------------------------|-----------------------------------|
|                      1 |             0.257 ± 0.002 seconds |
|                      4 |             0.137 ± 0.007 seconds |
|                      8 |             0.116 ± 0.006 seconds |

That's nearly double the throughput at higher concurrency levels.